requirements.txt
keras==2.8.0
numpy==1.22.3
psycopg2==2.9.3
psycopg2_binary==2.9.1
python-dotenv==0.20.0
scikit_learn==1.0.2
setuptools==57.1.0
tensorflow==2.8.0
